requirements.txt
mss>=9.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
pyperclip>=1.9.0
